Moved get_publisher_detail() from xss_client.py to xia_internal.py and applied isort
