+++ app/core/management/commands/extract_source_metadata.py
@@ -4,9 +4,9 @@ import logging
 from django.core.management.base import BaseCommand
 from django.utils import timezone
 
-from core.management.utils.xia_internal import get_source_metadata_key_value
+from core.management.utils.xia_internal import (get_publisher_detail,
+                                                get_source_metadata_key_value)
 from core.management.utils.xsr_client import read_source_file
-from core.management.utils.xss_client import get_publisher_detail
 from core.models import MetadataLedger
 
 logger = logging.getLogger('dict_config_logger')
