+++ app/core/management/commands/load_target_metadata.py
@@ -7,8 +7,8 @@ from django.core.serializers.json import DjangoJSONEncoder
 from django.db.models import Q
 from django.utils import timezone
 
+from core.management.utils.xia_internal import get_publisher_detail
 from core.management.utils.xis_client import response_from_xis
-from core.management.utils.xss_client import get_publisher_detail
 from core.models import MetadataLedger
 
 logger = logging.getLogger('dict_config_logger')
