+++ app/core/tests/test_commands_integration.py
@@ -6,7 +6,7 @@ from django.test import tag
 from django.utils import timezone
 
 from core.management.commands.extract_source_metadata import (
-    extract_metadata_using_key, get_publisher_detail, store_source_metadata)
+    extract_metadata_using_key, store_source_metadata)
 from core.management.commands.load_target_metadata import (
     post_data_to_xis, renaming_xia_for_posting_to_xis)
 from core.management.commands.transform_source_metadata import (
@@ -30,14 +30,6 @@ class CommandIntegration(TestSetUp):
     # globally accessible data sets
 
     # Test cases for extract_source_metadata
-
-    def test_get_publisher_detail(self):
-        """Test to get publisher name from XIAConfiguration """
-
-        xiaConfig = XIAConfiguration(publisher='JKO')
-        xiaConfig.save()
-        result_publisher = get_publisher_detail()
-        self.assertEqual('JKO', result_publisher)
 
     def test_store_source_metadata(self):
         """Test to extract data from Experience Source Repository(XSR)
